% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aav.R
\name{aav}
\alias{aav}
\title{Asymmetrical ANOVA From Beyond BACI and Post-Impact Designs}
\usage{
aav(data, aav.design, n.ftemp, n.fspac, names.impact, names.before = NULL,
  p.aav = 0.05, pooling = TRUE, method = "pool", p.pooling = 0.25,
  aci.tailed.test = 1, aav.file = NULL)
}
\arguments{
\item{data}{a data frame with temporal factors, spatial factors and variable(s).}

\item{aav.design}{\code{baci} for beyond before/after-control/impact design or \code{aci} for post-impact design.}

\item{n.ftemp}{number of temporal factors.}

\item{n.fspac}{number of spatial factors.}

\item{names.impact}{name of the asymmetrical level of the Control/Impact factor (e.g. "impact").}

\item{names.before}{only for \code{baci} design. Name of the before level of the Before/After factor (e.g. "before").}

\item{p.aav}{level of significance for the asymmetrical ANOVA. Default is \code{0.05}.}

\item{pooling}{remove a term from the analyse using a post-hoc pooling procedure. Default is \code{TRUE}.}

\item{method}{only for \code{baci} design and \code{pooling = TRUE}.
Method to be used for the pooling procedure.
\code{"pool"} to "pool" a term or a set of terms with zero contribution to the model or
\code{eliminate} to completely exclude a term from the model. Default is \code{"pool"}.}

\item{p.pooling}{only when \code{pooling = TRUE}.
Level of significance for the pooling procedure. Default is \code{0.25}.}

\item{aci.tailed.test}{only for \code{aci} design. Perform a 1-tailed or a 2-tailed test for all terms. Default is \code{1}.}

\item{aav.file}{A file path to save the AAV output files. Default \code{NULL} uses current name and working directory.}
}
\value{
\code{Excel} files containing the symmetrical ANOVA tables used to
run the asymmetrical ANOVA.
When analyzing more than one variable at once, each sheet have one ANOVA table.

\code{Excel} files containing the asymmetrical ANOVA table ("Complete Table" and "Final Table").
When analyzing more than one variable at once, each sheet have one ANOVA table.

A \code{list} containing the asymmetrical ANOVA table.
When analyzing more than one variable at once, each \code{sub-list} have one ANOVA table.
}
\description{
Calculate an asymmetrical ANOVA from a beyond BACI (before/after-control/impact)
design with multiple control locations, a post-impact design with only one
disturbed location and multiple control locations, or a post-impact design
with more than one place with one disturbed location and multiple control locations.

Asymmetrical ANOVAs are essentially multi-factorial.
Beyond BACI design can be used for models with 2 or 3 temporal factors
and 1 or 2 spatial factors. Post-impact design with only one disturbed location
can be used for models with 1 or 2 temporal factors
and 1 or 2 spatial factors. Post-impact design with more than one place with
one disturbed location can be used for models
with 2 or 3 spatial factors.
}
\details{
The construction of asymmetrical analyses of variance consists of combining
the Sum of Squares values from separate (symmetrical) ANOVAs.
In beyond BACI designs, four fully orthogonal ANOVAs need to be done: on all
the data, on "control" data, on "before" data, and on "control" and "before" data.
In post-impact designs with only one disturbed location two fully orthogonal
ANOVAs need to be done: on all the data and on "control" data.
In post-impact designs with more than one place with one disturbed location
two fully orthogonal and one fully nested ANOVAs need to be done:
on all data (orthogonal) and on "control" data (orthogonal and nested).
\code{aav} use \code{\link[GAD]{GAD}} (Sandrini-Neto et al. 2024) to perform symmetrical ANOVAs.

The \code{aav} analysis fits mixed asymmetrical ANOVA designs with a specific
combination of orthogonal/nested and fixed/random factors:
The  beyond BACI design runs for five-factors (3 temporal and 2 spatial),
four-factors (3 temporal and 1 spatial or 2 temporal and 2 spatial), and
three-factors (2 temporal and 1 spatial).
The post-impact design with only one disturbed location runs for
four-factors (2 temporal and 2 spatial), three-factors (2 temporal and 1 spatial or
1 temporal and 2 spatial), and two-factors (1 temporal and 1 spatial).
The post-impact design with more than one place with one disturbed location runs for
three-factors (0 temporal and 3 spatial) and two-factors (0 temporal and 2 spatial).

In all models the Location factor is the asymmetric one (e.g., one "impact"
group and two or more "control" groups), meaning there are different numbers
of levels of this nested factor within each different level of its upper-level
factor (e.g., Before vs After, Period, Time or Place factors).

In all models the temporal and spatial factors are orthogonal.
Within each factor type (temporal or spatial) the factors are hierarchically nested.
In the beyond BACI and post-impact design with only one disturbed location models,
the first factor and the Location factor are fixed and all others are random.
In the post-impact design with more than one place with one disturbed location,
the first factor is fixed and all other factors are random.

The data frame must follow a specific structure:
1- first the temporal factors, then the spatial factors, and then the variable(s);
2- each factor type must appear in a hierarchical order.

For beyond BACI design the order for temporal factors is Before_after, Period and Time (3 factors)
or Before_after and Time (2 factors).
The order for spatial factors is Location and Site (2 factors)
or just Location (1 factor).
For post-impact design with only one disturbed location and multiple control locations,
the order for temporal factors must be Period and Time (2 factors) or just Time (1 factor), and
the order for spatial factors must be Location and Site (2 factors) or just Location (1 factor).
For post-impact design with more than one place with one disturbed location and multiple control locations,
the order for spatial factors must be Place, Location and Site (3 factors) or Place and Location (2 factors).

If the names of the factors are different from the above names they will be changed.

When the data frame have more than one variable the univariate ANOVAs are run
for each one separated.

In the case of a pooling structure, both the \code{eliminate} and \code{"pool"} methods
are available for the beyond BACI design. For post-impact design when
\code{pooling} is \code{TRUE} the analysis is running using the method \code{"pool"}.

For \code{eliminate}, the degree of freedom (df) and sum of square (SS) for a
term is added to the residual df and SS and the term is treated as ever
having been part of the model. When \code{"pool"}, the df and SS for that term is
pooled with the term(s) that have an equivalent mean square expectation
after the component of variation of that term has been set to zero. This
component of variation is set to zero if it is >= \code{p.pooling} value.

\code{aav} "final table" is a typical ANOVA table ("Source.of.Variation","Df",
"Sum.Sq","Mean.Sq","F.value","P.Value.Upper.Tail","P.Value.Lower.Tail")
with the partitioning of the variance components between "control" and "impacts"
groups and between "before" and "after" disturbance groups (only for BACI designs).
The final ANOVA table has two additional step-by-step elements:
a "Next.Step" that indicates where to go from here, according to the
calculation of the mean square estimates for that term,
and an "Interpretation" that states the results found.
There is no "P.Value.Lower.Tail" for the a post-impact design with more than
one place with one disturbed location. For beyond BACI design "P.Value.Lower.Tail"
appears only when their is a valid test for a two-tailed analysis;

The "complete table" presents all elements used in the asymmetrical analysis,
with the "final table" elements plus:
The complement for a two-tailed analysis ("P.Value.Lower.Tail");
the degree of freedom formula ("Df.Formula");
from which symmetrical ANOVA derived the SS and Df values ("Calculated.From");
who is the denominator to the calculus of the F-ratio for that term ("Denominator");
the Df, SS and MS for the pooled term(s) ("Df.Pooled","Sum.Sq.Pooled","Mean.Sq.Pooled");
based on the \code{p.aav} and \code{p.pooling} values, a statement if that term is significant,
non-significant or can be eliminated ("Significant");
based on the \code{p.pooling} value and the mean square estimates, a statement on
which of the possible pooled term were used ("Post.Hoc.Pooling");
some designs has two-tailed analysis ("P.Value.Lower.Tail").
}
\examples{
# Example for BACI design containing 3 temporal factors and 2 spatial factor:
data(baci.3t2s)
ex.baci.3t2s<-aav(data = baci.3t2s, aav.design = "baci", n.ftemp = 3, n.fspac = 2,
    names.impact = "impact", names.before = "before",
    pooling = TRUE, method = "eliminate", p.aav = 0.05, p.pooling = 0.05)
ex.baci.3t2s

# Example for post-impact design with only one disturbed location
# containing 2 temporal factors and 2 spatial factor:
data(aci.2t2s)
ex.aci.2t2s<-aav(data = aci.2t2s, aav.design = "aci", n.ftemp = 2, n.fspac = 2,
   names.impact = "impact", pooling = TRUE, p.aav = 0.05, p.pooling = 0.25)
ex.aci.2t2s

# Example for post-impact design with more than one place with one
# disturbed location containing 0 temporal factors and 3 spatial factor:
 data(aci.0t3s)
ex.aci.0t3s<-aav(data = aci.0t3s, aav.design = "aci", n.ftemp = 0, n.fspac = 3,
    names.impact = "impact", pooling = TRUE, p.aav = 0.05, p.pooling = 0.25)
ex.aci.0t3s

}
\references{
Glasby, T.M. 1997. Analysing data from post-impact studies using asymmetrical analyses
of variance: A case study of epibiota on marinas. Aust. J. Ecol., 22, 448-459.
\doi{10.1111/j.1442-9993.1997.tb00696.x}.

Sandrini-Neto, L., Gilbert, E., Camargo, M. 2024. \emph{GAD: Analysis of Variance from General Principles}.
R package version 2.0, \doi{10.32614/CRAN.package.GAD}.

Underwood, A.J. 1993. The mechanics of spatially replicated sampling programmes
to detect environmental impacts in a variable world. Aust. J. Ecol. 18,99-116.
\doi{10.1111/j.1442-9993.1993.tb00437.x}.

Underwood, A.J. 1997. \emph{Experiments in Ecology: Their Logical Design and Interpretation Using
Analysis of Variance}. Cambridge University Press, Cambridge. \doi{10.1017/CBO9780511806407}.
}
\seealso{
\code{\link[AAV]{sym4asym}},
\code{\link[AAV]{asym.aci}},
\code{\link[AAV]{asym.baci}},
\code{\link[GAD]{gad}}
}
\author{
Paulo Pagliosa \email{paulo.pagliosa@ufsc.br}
}
