% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aci.0t3s}
\alias{aci.0t3s}
\title{Dataset For Testing a Post-Impact Design
With More Than One Place With One Disturbed Location and Multiple
Control Locations}
\format{
A dataframe with 288 rows and 6 columns.
}
\source{
Created in-house to serve as an example.
}
\usage{
aci.0t3s
}
\description{
Dataset with 0 temporal and 3 spatial (Place, Location, Site) factors
and 3 response variables simulating no impact, press disturbance
and pulse disturbance.
}
\details{
\itemize{
\item Place: a fixed factor with two levels (pl1, pl2).
\item Location: a random factor with three levels (control1, control2, impact) nested in Places.
\item Site: a random factor with two levels (s1, s2) nested in Location.
\item Var.random: response variable with random numbers; no impact simulation.
\item Var.L.impact: response variable; press impact - a reduction in numbers at the impact location is simulated.
\item Var.S.impact: response variable; pulse impact - a reduction in numbers at one site of the impact location is simulated.
}
}
\examples{
 data(aci.0t3s)
 aci.0t3s
}
\keyword{datasets}
