% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aci.2t2s}
\alias{aci.2t2s}
\title{Dataset For testing a Post-Impact Design
With Only One Disturbed Location and Multiple Control Locations}
\format{
A dataframe with 288 rows and 7 collumns.
}
\source{
Created in-house to serve as an example.
}
\usage{
aci.2t2s
}
\description{
Dataset with 2 temporal (Period, Time) and
2 spatial (Location, Site) factors and 3 response variables simulating
no impact, press disturbance and pulse disturbance.
}
\details{
\itemize{
\item Period: a fixed factor with two levels (p1, p2).
\item Time: a random factor with two levels (t1, t2) nested in Period.
\item Location: a random factor with three levels (control1, control2, impact).
\item Site: a random factor with two levels (s1, s2) nested in Location.
\item Var.random: response variable with random numbers; no impact simulation.
\item Var.L.impact: response variable; press impact - a reduction in numbers at the impact location is simulated.
\item Var.S.impact: response variable; pulse impact - a reduction in numbers at one site of the impact location is simulated.
}
}
\examples{
 data(aci.2t2s)
 aci.2t2s
}
\keyword{datasets}
