% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asym.aci.R
\name{asym.aci}
\alias{asym.aci}
\title{Asymmetrical ANOVA From Post-Impact Design}
\usage{
asym.aci(data, n.ftemp, n.fspac, names.impact, sym.F, sym.C, sym.NC = NULL,
  p.aav = 0.05, pooling = TRUE, p.pooling = 0.25, aci.tailed.test = 1)
}
\arguments{
\item{data}{a data frame with temporal factors and spatial factors.}

\item{n.ftemp}{number of temporal factors.}

\item{n.fspac}{number of spatial factors.}

\item{names.impact}{name of the asymmetrical level of the Control/Impact factor (e.g. "impact").}

\item{sym.F}{an object containing the results returned by \code{sym4asym} function or
an symmetrical analysis of variance table run with all the data,
meaning the asymmetry is deal as a factor level (e.g., Control_1,
Control_2 and Impact are the factor levels).}

\item{sym.C}{an object containing the results returned by \code{sym4asym} function or
an symmetrical analysis of variance table run omitting the asymmetrical data
(e.g., only the Control_1 and Control_2 are the factor levels).}

\item{sym.NC}{only for a design with more than one place with one disturbed location.
an object containing the results returned by \code{sym4asym} function or
an symmetrical analysis of variance table run omitting the asymmetrical data
and with all factors nested (e.g., Control location nested in Place).}

\item{p.aav}{level of significance for the asymmetrical ANOVA. Default is \code{0.05}.}

\item{pooling}{remove a term from the analyse using a post-hoc pooling procedure. Default is \code{TRUE}.}

\item{p.pooling}{only when \code{pooling = TRUE}.
Level of significance for the pooling procedure. Default is \code{0.25}.}

\item{aci.tailed.test}{Perform a 1-tailed or a 2-tailed test for all terms. Default is \code{1}.}
}
\value{
A \code{data frame} containing an asymmetrical ANOVA table.
}
\description{
Calculate an asymmetrical ANOVA from post-impact design with only one
disturbed location and multiple control locations or
a post-impact design with more than one place with one
disturbed location and multiple control.

Post-impact design with only one disturbed location can be used for models
with 1 or 2 temporal factors and 1 or 2 spatial factors.
Post-impact design with more than one place with one disturbed location
can be used for models with 2 or 3 spatial factors.
}
\details{
The construction of asymmetrical analyses of variance consists of combining
the Sum of Squares values from separate (symmetrical) ANOVAs.
In post-impact designs with only one disturbed location two fully orthogonal
ANOVAs need to be done: on all the data and on "control" data.
In post-impact designs with more than one place with one disturbed location
two fully orthogonal and one fully nested ANOVAs need to be done:
on all data (orthogonal) and on "control" data (orthogonal and nested).

The \code{asym.aci} analysis fits mixed ANOVA designs with a specific
combination of orthogonal/nested and fixed/random factors:
The post-impact design with only one disturbed location runs for
four-factors (2 temporal and 2 spatial), three-factors (2 temporal and 1 spatial or
1 temporal and 2 spatial), and two-factors (1 temporal and 1 spatial).
The post-impact design with more than one place with one disturbed location runs for
three-factors (0 temporal and 3 spatial) and two-factors (0 temporal and 2 spatial).

In all models the Location factor is the asymmetric one (e.g., one "impact"
group and two or more "control" groups), meaning there are different numbers
of levels of this nested factor within each different level of its upper-level
factor (e.g., Period, Time or Place factors).

In all models the temporal and spatial factors are orthogonal.
Within each factor type (temporal or spatial) the factors are hierarchically nested.
In the post-impact design with only one disturbed location models,
the first factor and the Location factor are fixed and all others are random.
In the post-impact design with more than one place with one disturbed location,
the first factor is fixed and all other factors are random.

The data frame must follow a specific structure:
1- first the temporal factors, then the spatial factors;
2- each factor type must appear in a hierarchical order.

For post-impact design with only one disturbed location and multiple control locations,
the order for temporal factors must be Period and Time (2 factors) or just Time (1 factor), and
the order for spatial factors must be Location and Site (2 factors) or just Location (1 factor).
For post-impact design with more than one place with one disturbed location and multiple control locations,
the order for spatial factors must be Place, Location and Site (3 factors) or Place and Location (2 factors).

If the names of the factors are different from the above names they will be changed.

In the case of a pooling structure, when \code{pooling} is set to \code{TRUE},
the analysis runs using the \code{pool} method, where the
degrees of freedom (df) and sums of squares (SS) for a given term are
pooled with the term(s) that have an equivalent mean square expectation
after the component of variation of that term has been set to zero. This
component of variation is set to zero if it is >= \code{p.pooling} value.

\code{asym.aci} "final table" is a typical ANOVA table ("Source.of.Variation","Df",
"Sum.Sq","Mean.Sq","F.value","P.Value.Upper.Tail","P.Value.Lower.Tail") with the
partitioning of the variance components between "control" and "impacts" groups.
The final ANOVA table has two additional step-by-step elements:
a "Next.Step" that indicates where to go from here, according to the
calculation of the mean square estimates for that term,
and an "Interpretation" that states the results found.
There is no "P.Value.Lower.Tail" for the a post-impact design with more than
one place with one disturbed location.

The "complete table" presents all elements used in the asymmetrical analysis,
with the "final table" elements plus:
the degree of freedom formula ("Df.Formula");
from which symmetrical ANOVA derived the SS and Df values ("Calculated.From");
who is the denominator to the calculus of the F-ratio for that term ("Denominator");
the Df, SS and MS for the pooled term(s) ("Df.Pooled","Sum.Sq.Pooled","Mean.Sq.Pooled");
based on the \code{p.aav} and \code{p.pooling} values, a statement if that term is significant,
non-significant or can be eliminated ("Significant");
based on the \code{p.pooling} value and the mean square estimates, a statement on
which of the possible pooled term were used ("Post.Hoc.Pooling").
}
\examples{
# Example for post-impact design with only one disturbed location
# containing 2 temporal factors and 2 spatial factor:
#library(AAV)
data(aci.2t2s) # 4-factors dataset
data(symANOVA_aci.2t2s) # symmetrical ANOVAS from aci.2t2s data
ex.symANOVA_aci.2t2s<-asym.aci(data = aci.2t2s, n.ftemp = 2, n.fspac = 2,
   names.impact = "impact",
   sym.F = symANOVA_aci.2t2s$full_design,
   sym.C = symANOVA_aci.2t2s$control_design,
   p.aav = 0.05, pooling = TRUE, p.pooling = 0.25, aci.tailed.test = 1)
ex.symANOVA_aci.2t2s

# Example for post-impact design with more than one place with one
# disturbed location containing 0 temporal factors and 3 spatial factor:
#library(AAV)
data(aci.0t3s) # 3-factors dataset
data(symANOVA_aci.0t3s) # symmetrical ANOVAS from aci.0t3s data
ex.symANOVA_aci.0t3s<-asym.aci(data = aci.0t3s, n.ftemp = 0, n.fspac = 3,
    names.impact = "impact",
    sym.F = symANOVA_aci.0t3s$full_design,
    sym.C = symANOVA_aci.0t3s$control_design,
    sym.NC = symANOVA_aci.0t3s$nested_control_design,
    p.aav = 0.05, pooling = TRUE, p.pooling = 0.25)
ex.symANOVA_aci.0t3s

}
\references{
Glasby, T.M. 1997. Analysing data from post-impact studies using asymmetrical analyses
of variance: A case study of epibiota on marinas. Aust. J. Ecol., 22, 448-459.
\doi{10.1111/j.1442-9993.1997.tb00696.x}.

Underwood, A.J. 1993. The mechanics of spatially replicated sampling programmes
to detect environmental impacts in a variable world. Aust. J. Ecol. 18,99-116.
\doi{10.1111/j.1442-9993.1993.tb00437.x}.

Underwood, A.J. 1997. \emph{Experiments in Ecology: Their Logical Design and Interpretation Using
Analysis of Variance}. Cambridge University Press, Cambridge. \doi{10.1017/CBO9780511806407}.
}
\seealso{
\code{\link[AAV]{sym4asym}},
\code{\link[AAV]{asym.baci}},
\code{\link[AAV]{aav}}
}
\author{
Paulo Pagliosa \email{paulo.pagliosa@ufsc.br}
}
