% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baci.3t2s}
\alias{baci.3t2s}
\title{Dataset For Testing a Beyond BACI Design}
\format{
A dataframe with 288 rows and 8 columns.
}
\source{
Created in-house to serve as an example.
}
\usage{
baci.3t2s
}
\description{
Dataset with 3 temporal (Before-after, Period, Time) and
2 spatial (Location, Site) factors and 3 response variables simulating
no impact, press disturbance and pulse disturbance.
}
\details{
\itemize{
\item Before-after: a fixed factor with two levels (before, after).
\item Period: a random factor with two levels (p1, p2) nested in Before-after.
\item Time: a random factor with two levels (t1, t2) nested in Period.
\item Location: a random factor with three levels (control1, control2, impact).
\item Site: a random factor with two levels (s1, s2) nested in Location.
\item Var.random: response variable with random numbers; no impact simulation.
\item Var.L.impact: response variable; press impact - a reduction in numbers at the impact location is simulated.
\item Var.S.impact: response variable; pulse impact - a reduction in numbers at one site of the impact location is simulated.
}
}
\examples{
 data(baci.3t2s)
 baci.3t2s
}
\keyword{datasets}
