% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmagnitude.R
\name{ggmagnitude}
\alias{ggmagnitude}
\title{Create a plot showing the estimated magnitude of an ecological impact}
\usage{
ggmagnitude(data, aav.design, n.ftemp, n.fspac)
}
\arguments{
\item{data}{a list genereted from \code{\link[AAV]{magnitude}}.}

\item{aav.design}{\code{baci} for beyond Before/After-Control/Impact design or \code{aci} for Post-Impact design.}

\item{n.ftemp}{number of temporal factors.}

\item{n.fspac}{number of spatial factors.}
}
\value{
The function returns a xy plot with the original (top right), a random (= 0; bottom left)
and the estimated alterations on the mean and variance of the data (center).
}
\description{
It provides a customised 'ggplot' for displaying the magnitude of an ecological
impact based on the results from \code{\link[AAV]{magnitude}}.
}
\details{
The magnitude of an ecological impact is estimated using the same data types, designs, factors
and parameters as in \code{\link[AAV]{aav}}, but only considers one variable at a time.

The estimates are plotted to each relevant term from the asymmetrical analysis of variance
in order to determine the magnitude of the impact.
The terms can be interpreted separated each other.

To each term there is a chart showing the observed p-values from the original data,
and the expected p-value from a random data (resampled the original data)
and from each estimated alteration on the mean and variance of the data.
The magnitude of the impact can be discovered when the original p-value from a term
is significant (p < 0.05). In this case, the first estimate of the altered mean and variance
that is significant indicate the minimum magnitude of the impact that the original dataset can detect.

Wheather a term is significant, it could detect:

In baci desing:
\itemize{
\item Before vs After x Impact vs Controls = B x I:   Long-term local-scale impact
\item B x Site(I) = B x S(I):   Long-term small-scale impact
\item Perido(After) x I = P(Aft) x I:   Medium-term local-scale impact
\item Time(P(Aft)) x I = T(P(Aft)) x I:   Medium-term small-scale impact
\item P(Aft) x S(I) or T(Aft) x I:   Short-term local-scale impact
\item T(P(Aft)) x S(I) or T(Aft) x S(I):   Short-term local-scale impact
}

In post-impact design with only one disturbed location and multiple control locations:
\itemize{
\item Impact vs Controls = I:   Long-term local-scale impact
\item Site(I) = S(I):   Long-term small-scale impact
\item Period x I = P x I:   Medium-term local-scale impact
\item P x S(I):   Medium-term small-scale impact
\item T(P) x I or T x I:   Short-term local-scale impact
\item T(P) x S(I) or T x S(I):   Short-term and small-scale impact
}

In post-impact design involving more than one place with one disturbed
location and multiple control locations:
\itemize{
\item Treatment = T:   Generalized local-scale impact
\item Site(T) = S(T):   Generalized site-scale impact
\item Place x T = P x T:   Local-specific scale impact
\item P x S(T):   Site-specific scale impact
}
}
\examples{
# Example for BACI design containing 2 temporal factors and 1 spatial factor.
# using the vectors:
# magnitude_md = c(0.10, 0.20, 0.30, 0.40, 0.50, 0.60, 0.70, 0.80, 0.90, 1) and
# magnitude_var = c(0.1, 1, 2, 5, 10, 25, 50, 75, 100)
library(ggplot2)
data(ex.ggmag)
ggmagnitude(data = ex.ggmag, aav.design = "baci", n.ftemp = 2, n.fspac = 1)

}
\references{
Underwood, A.J. 1997. \emph{Experiments in Ecology: Their Logical Design and Interpretation Using
Analysis of Variance}. Cambridge University Press, Cambridge. \doi{10.1017/CBO9780511806407}.
}
\seealso{
\code{\link[AAV]{aav}},
\code{\link[GAD]{magnitude}}
}
\author{
Paulo Pagliosa \email{paulo.pagliosa@ufsc.br}
}
