% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnitude.R
\name{magnitude}
\alias{magnitude}
\title{Estimates the magnitude of an ecological impact}
\usage{
magnitude(data, aav.design, n.ftemp, n.fspac, names.impact,
  names.before = NULL, p.aav = 0.05, pooling = TRUE, method = "pool",
  p.pooling = 0.25, aci.tailed.test = 1, magnitude_md = c(0.1, 0.2, 0.3,
  0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1), magnitude_var = c(0.1, 1, 2, 5, 10, 25,
  50, 75, 100))
}
\arguments{
\item{data}{a data frame with temporal factors, spatial factors and variable(s).}

\item{aav.design}{\code{baci} for beyond Before/After-Control/Impact design or \code{aci} for Post-Impact design.}

\item{n.ftemp}{number of temporal factors.}

\item{n.fspac}{number of spatial factors.}

\item{names.impact}{name of the asymmetrical level of the Location factor (e.g. 'impact').}

\item{names.before}{only for \code{baci} design. Name of the before level of the Before/After factor (e.g. 'before').}

\item{p.aav}{level of significance for the asymmetrical ANOVA. Default is \code{0.05}.}

\item{pooling}{remove a term from the analysis using a post-hoc pooling procedure. Default is \code{TRUE}.}

\item{method}{only when \code{baci} design and \code{pooling = TRUE}.
Method to be used for the pooling procedure.
\code{pool} to pool a term or a set of terms with zero contribution to the model or
\code{eliminate} to completely exclude a term from the model. Default is \code{pool}.}

\item{p.pooling}{only when \code{pooling = TRUE}.
Level of significance for the pooling procedure. Default is \code{0.25}.}

\item{aci.tailed.test}{only for \code{aci} post-impact design with only one disturbed location
and multiple control locations. Perform a 1-tailed or a 2-tailed test for all terms.
Default is \code{1}.}

\item{magnitude_md}{a vector to estimate alterations to the mean values of the data.}

\item{magnitude_var}{a vector to estimate alterations to the variance values of the data.}
}
\value{
The function returns a list containing the following items:
\item{data_magimp}{the original dataset, along with the simulated  magnitude
of the impact on the mean and variance of the data.}
\item{aav_test}{the result of the complete asymmetrical ANOVA tables for the original
dataset and for each simulation of the magnitude of the impact on the mean and variance of the data.}
\item{extract_aav}{a table with the characterization of the design and factors used
('aav.design','n.ftemp','n.fspac'),
a code for the type of the impact ('impType'), the magnitude of the variance alteration ('magVAR'),
the magnitude of the mean alteration ('magMD'), the corresponding term of the analysis ('term'),
and the P-Value from the specific analysis ('p.value').}
}
\description{
It provides an estimate of the magnitude of the impact on the mean and variance of the data
that could be detected, given the observed P-Value from the asymmetrical analysis of variance.
}
\details{
The magnitude of the impact is estimated using the same data types, designs, factors
and parameters as in \code{\link[AAV]{aav}}, but only considers one variable at a time.

The estimates are made to each relevant term from the asymmetrical analysis of variance
in order to determine the magnitude of the impact.
The terms can be interpreted separated each other.
The relevant terms are:

In baci desing:
\itemize{
\item Before vs After x Impact vs Controls = B x I:   Long-term local-scale impact
\item B x Site(I) = B x S(I):   Long-term small-scale impact
\item Perido(After) x I = P(Aft) x I:   Medium-term local-scale impact
\item Time(P(Aft)) x I = T(P(Aft)) x I:   Medium-term small-scale impact
\item P(Aft) x S(I) or T(Aft) x I:   Short-term local-scale impact
\item T(P(Aft)) x S(I) or T(Aft) x S(I):   Short-term local-scale impact
}

In post-impact design with only one disturbed location and multiple control locations:
\itemize{
\item Impact vc Controls = I:   Long-term local-scale impact
\item Site(I) = S(I):   Long-term small-scale impact
\item Period x I = P x I:   Medium-term local-scale impact
\item P x S(I):   Medium-term small-scale impact
\item T(P) x I or T x I:   Short-term local-scale impact
\item T(P) x S(I) or T x S(I):   Short-term and small-scale impact
}

In post-impact design involving more than one place with one disturbed
location and multiple control locations:
\itemize{
\item Treatment = T:   Generalized local-scale impact
\item Site(T) = S(T):   Generalized site-scale impact
\item Place x T = P x T:   Local-specific scale impact
\item P x S(T):   Site-specific scale impact
}
}
\examples{
# Example for BACI design containing 3 temporal factors and 2 spatial factor:
data(baci.3t2s)
ex.baci.3t2s<-magnitude(data = baci.3t2s, aav.design = "baci", n.ftemp = 3, n.fspac = 2,
    names.impact = "impact", names.before = "before",
    pooling = TRUE, method = "eliminate", p.aav = 0.05, p.pooling = 0.05,
    magnitude_md = c(0.50, 1),
    magnitude_var = c(0.1, 1))
ex.baci.3t2s

}
\references{
Underwood, A.J. 1997. \emph{Experiments in Ecology: Their Logical Design and Interpretation Using
Analysis of Variance}. Cambridge University Press, Cambridge. \doi{10.1017/CBO9780511806407}.
}
\seealso{
\code{\link[AAV]{aav}},
\code{\link[GAD]{ggmagnitude}}
}
\author{
Paulo Pagliosa \email{paulo.pagliosa@ufsc.br}
}
