% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{symANOVA_aci.0t3s}
\alias{symANOVA_aci.0t3s}
\title{ANOVA Tables for a Post-Impact Design With More Than One Place With One Disturbed Location and Multiple Control Locations}
\format{
A list with 3 sub lists.
}
\source{
Created in-house to serve as an example.
}
\usage{
symANOVA_aci.0t3s
}
\description{
There are three analysis of variance tables: one based on the full dataset (orthogonal),
and two based on the dataset of only the control locations (one is orthogonal and the other is nested).
Analysis of variance table from 0 temporal factors and 3 spatial factors (D, E and G).
Source of variation codes in AAV are A (Before-after), B (Period), C (Time), D (Place), E (Location), G (Site).
}
\details{
The ANOVAS are from 0 temporal factors and 3 spatial factors (D, E and G).
Source of variation codes in AAV are A (Before-after), B (Period), C (Time), D (Place), E (Location), G (Site).
Each table is provided with:
\itemize{
\item Source.of.variation: Source of variation.
\item Df: Degree of freedom.
\item Sum.Sq: Sum of squares.
\item Mean.Sq: Mean square.
\item F.value: F-test.
\item Pr(>F): P-value.
}
}
\examples{
 data(symANOVA_aci.0t3s)
 symANOVA_aci.0t3s
}
\keyword{datasets}
