% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{symANOVA_aci.2t2s}
\alias{symANOVA_aci.2t2s}
\title{ANOVA Tables for a Post-Impact Design With Only One Disturbed Location and Multiple Control Locations}
\format{
A list with 2 sub lists.
}
\source{
Created in-house to serve as an example.
}
\usage{
symANOVA_aci.2t2s
}
\description{
There are two analysis of variance tables: one with the full dataset,
and one with only the control locations dataset.
Analysis of variance table from 2 temporal factors (B and C) and 2 spatial factors (E and G).
Source of variation codes in AAV are A (Before-after), B (Period), C (Time), D (Place), E (Location), G (Site).
}
\details{
The ANOVAS are from 2 temporal factors (B and C) and 2 spatial factors (E and G).
Source of variation codes in AAV are A (Before-after), B (Period), C (Time), D (Place), E (Location), G (Site).
Each table is provided with:
\itemize{
\item Source.of.variation: Source of variation.
\item Df: Degree of freedom.
\item Sum.Sq: Sum of squares.
\item Mean.Sq: Mean square.
\item F.value: F-test.
\item Pr(>F): P-value.
}
}
\examples{
 data(symANOVA_aci.2t2s)
 symANOVA_aci.2t2s
}
\seealso{
\code{\link[AAV]{aav}},
\code{\link[AAV]{asym.aci}},
\code{\link[AAV]{aci.2t2s}}
}
\keyword{datasets}
