% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{symANOVA_baci.3t2s}
\alias{symANOVA_baci.3t2s}
\title{ANOVA Tables for a Beyond BACI Design}
\format{
A list with 4 sub lists.
}
\source{
Created in-house to serve as an example.
}
\usage{
symANOVA_baci.3t2s
}
\description{
There are four analysis of variance tables: one with the full dataset;
one with the dataset from the control locations only;
one with the data from before the disturbance occurred only;
and one with the dataset from the control locations only from before the disturbance occurred.
The ANOVAS are from 3 temporal factors (A, B and C) and 2 spatial factors (E and G).
Source of variation codes in AAV are A (Before-after), B (Period), C (Time), D (Place), E (Location), G (Site).
}
\details{
The ANOVAS are from 3 temporal factors (A, B and C) and 2 spatial factors (E and G).
Source of variation codes in AAV are A (Before-after), B (Period), C (Time), D (Place), E (Location), G (Site).
Each table is provided with:
\itemize{
\item Source.of.variation: Source of variation.
\item Df: Degree of freedom.
\item Sum.Sq: Sum of squares.
\item Mean.Sq: Mean square.
\item F.value: F-test.
\item Pr(>F): P-value.
}
}
\examples{
 data(symANOVA_baci.3t2s)
 symANOVA_baci.3t2s
}
\keyword{datasets}
