% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asym.aci.R
\name{asym.aci}
\alias{asym.aci}
\title{Asymmetrical ANOVA From Post-Impact Design}
\usage{
asym.aci(data, n.ftemp, n.fspac, names.impact, sym.F, sym.C, sym.NC = NULL,
  p.aav = 0.05, pooling = TRUE, p.pooling = 0.25, aci.tailed.test = 1)
}
\arguments{
\item{data}{a data frame with temporal factors, spatial factors and variable(s).}

\item{n.ftemp}{number of temporal factors.}

\item{n.fspac}{number of spatial factors.}

\item{names.impact}{name of the asymmetrical level of the Location factor (e.g. 'impact').}

\item{sym.F}{an object containing the results returned by \code{sym4asym} function or
an symmetrical analysis of variance table run with all the data,
meaning the asymmetry is deal as a factor level (e.g., Control_1,
Control_2 and Impact are the factor levels).}

\item{sym.C}{an object containing the results returned by \code{sym4asym} function or
an symmetrical analysis of variance table run omitting the asymmetrical data
(e.g., only the Control_1 and Control_2 are the factor levels).}

\item{sym.NC}{only for a design with more than one place with one disturbed location.
An object containing the results returned by \code{sym4asym} function or
an symmetrical analysis of variance table run omitting the asymmetrical data
and with all factors nested (e.g., Control location nested in Place).}

\item{p.aav}{level of significance for the asymmetrical ANOVA. Default is \code{0.05}.}

\item{pooling}{remove a term from the analysis using a post-hoc pooling procedure. Default is \code{TRUE}.}

\item{p.pooling}{only when \code{pooling = TRUE}.
Level of significance for the pooling procedure. Default is \code{0.25}.}

\item{aci.tailed.test}{only for post-impact design with only one disturbed location
and multiple control locations. Perform a 1-tailed or a 2-tailed test for all terms.
Default is \code{1}.}
}
\value{
The function returns a list containing the following items:
\item{sym.anova}{the symmetrical ANOVA tables used for running asymmetrical ANOVAs.}
\item{asym.anova_final}{the final asymmetrical ANOVA tables.
The 'final table' is a typical ANOVA table
('Source.of.Variation','Df', 'Sum.Sq','Mean.Sq','F.value','P.Value.Upper.Tail','P.Value.Lower.Tail')
with the partitioning of the variance components between control and impacts
groups and between before and after disturbance groups (only for BACI designs).
The 'final table' has two additional step-by-step elements:
a 'Next.Step' that indicates where to go from here, according to the
calculation of the mean square estimates for that term,
and an 'Interpretation' that states the results found.
There is no 'P.Value.Lower.Tail' for the a post-impact design with more than
one place with one disturbed location. For beyond BACI design 'P.Value.Lower.Tail'
appears only when there is a valid test for a two-tailed analysis.}
\item{asym.anova_complete}{the complete asymmetrical ANOVA tables.
The 'complete table' presents all elements used in the asymmetrical analysis,
with the 'final table' elements plus:
The complement for a two-tailed analysis ('P.Value.Lower.Tail');
the degrees of freedom formula ('Df.Formula');
from which symmetrical ANOVA derived the SS and Df values ('Calculated.From');
who is the denominator to the calculus of the F-ratio for that term ('Denominator');
the Df, SS and MS for the pooled term(s) ('Df.Pooled','Sum.Sq.Pooled','Mean.Sq.Pooled');
based on the \code{p.aav} and \code{p.pooling} values, a statement if that term is significant,
non-significant or can be eliminated ('Significant');
based on the \code{p.pooling} value and the mean square estimates, a statement on
which of the possible pooled term were used ('Post.Hoc.Pooling');
some designs has two-tailed analysis ('P.Value.Lower.Tail').}

When analysing more than one variable at once, the results for each variable
appear in the sublists.
}
\description{
Provides an asymmetrical ANOVA from post-impact design with only one
disturbed location and multiple control locations or
a post-impact design involving more than one place with one
disturbed location and multiple control.

Post-impact design with only one disturbed location can be used for models
with 1 or 2 temporal factors and 1 or 2 spatial factors.
Post-impact design involving more than one place with one disturbed location
can be used for models with 2 or 3 spatial factors.
}
\details{
Asymmetrical analyses of variance are constructed by combining
the Sum of Squares values from separate (symmetrical) ANOVAs.
In post-impact designs with only one disturbed location two fully orthogonal
ANOVAs are performed: one on all the data, and one on the 'control' data.
In post-impact designs involving more than one place with one disturbed location,
two fully orthogonal and one fully nested ANOVA need to be performed: one on all the data
(orthogonal) and two on the 'control' data (orthogonal and nested).
\code{aav} uses \code{\link[GAD]{GAD}} (Sandrini-Neto et al. 2024) to perform symmetrical ANOVAs.

The \code{asym.aci} analysis fits mixed asymmetrical ANOVA designs with a specific
combination of orthogonal/nested and fixed/random factors:
The post-impact design with only one disturbed location runs for
four factors (2 temporal and 2 spatial), three factors (2 temporal and 1 spatial or
1 temporal and 2 spatial), and two factors (1 temporal and 1 spatial).
The post-impact design involving more than one place with one disturbed location runs for
three factors (0 temporal and 3 spatial) and two factors (0 temporal and 2 spatial).

In all models, the Location factor is asymmetric (e.g., one 'impact' group and
two or more 'control' groups), meaning there are different numbers of levels of this
nested factor at each  level of its upper-level factor (e.g., Place factor) or
contrasting factors (e.g., Before-After, Period, Time).

In all models the temporal and spatial factors are orthogonal.
Within each factor type (temporal or spatial) the factors are hierarchically nested.
In the post-impact design with only one disturbed location models,
the first factor and the Location factor are fixed and all others are random.
In the post-impact design involving than one place with one disturbed location,
the first factor is fixed and the others are random.

The data frame must follow a specific structure:
1- first the temporal factors, then the spatial factors, and then the variable(s);
2- each factor type must appear in a hierarchical order.

In a post-impact design involving only one disturbed location and multiple control locations,
the temporal factor order must be either Period and Time (2 factors) or just Time (1 factor), and
the spatial factor order must be either Location and Site (2 factors) or just Location (1 factor).
In a post-impact design involving more than one place with one disturbed location and multiple control locations,
the order of the spatial factors must be Place, Location and Site (3 factors) or Place and Location (2 factors).

If the names of the factors do not match those listed above, \code{AAV} will change them.

When the data frame contains more than one variable, separate univariate ANOVAs are run for each one.

In the case of a pooling structure, both the \code{pool} and \code{eliminate} methods
are available for the beyond BACI design. For post-impact design when
\code{pooling} is \code{TRUE} the analysis is running using the method \code{pool}.

In the case of a pooling structure, when \code{pooling = TRUE},
the analysis runs using the \code{pool} method, where the
degrees of freedom (Df) and sums of squares (SS) for that term are
pooled with the term(s) that have an equivalent mean square expectation
after the component of variation of that term has been set to zero. This
component of variation is set to zero if it is >= \code{p.pooling} value.

The magnitude of interactions in time and space can be analysed using a two-tailed tests.
for Post-Impact design involving only one impact location,
where temporal interactions among locations occur just after the disturbance,
a two-tailed test can be performed (by selecting \code{aci.tailed.test = 2}) to determine
whether and in which direction the time trends differ between impact and controls locations.
For a Post-Impact design involving more than one place with one disturbed location,
there is no need for a two-tailed test. In this design, F-tests contrast treatments
against an interaction term of spatial scale and treatments and are properly run as one-tailed tests.
}
\examples{
# Example for post-impact design with only one disturbed location
# containing 2 temporal factors and 2 spatial factor:
data(aci.2t2s) # 4-factors dataset
data(symANOVA_aci.2t2s) # symmetrical ANOVAS from aci.2t2s data
ex.symANOVA_aci.2t2s<-asym.aci(data = aci.2t2s, n.ftemp = 2, n.fspac = 2,
   names.impact = "impact",
   sym.F = symANOVA_aci.2t2s$full_design,
   sym.C = symANOVA_aci.2t2s$control_design,
   p.aav = 0.05, pooling = TRUE, p.pooling = 0.25, aci.tailed.test = 1)
ex.symANOVA_aci.2t2s

# Example for post-impact design with more than one place with one
# disturbed location containing 0 temporal factors and 3 spatial factor:
data(aci.0t3s) # 3-factors dataset
data(symANOVA_aci.0t3s) # symmetrical ANOVAS from aci.0t3s data
ex.symANOVA_aci.0t3s<-asym.aci(data = aci.0t3s, n.ftemp = 0, n.fspac = 3,
    names.impact = "impact",
    sym.F = symANOVA_aci.0t3s$full_design,
    sym.C = symANOVA_aci.0t3s$control_design,
    sym.NC = symANOVA_aci.0t3s$nested_control_design,
    p.aav = 0.05, pooling = TRUE, p.pooling = 0.25)
ex.symANOVA_aci.0t3s

}
\references{
Glasby, T.M. 1997. Analysing data from post-impact studies using asymmetrical analyses
of variance: A case study of epibiota on marinas. Aust. J. Ecol., 22, 448-459.
\doi{10.1111/j.1442-9993.1997.tb00696.x}.

Underwood, A.J. 1993. The mechanics of spatially replicated sampling programmes
to detect environmental impacts in a variable world. Aust. J. Ecol. 18,99-116.
\doi{10.1111/j.1442-9993.1993.tb00437.x}.

Underwood, A.J. 1997. \emph{Experiments in Ecology: Their Logical Design and Interpretation Using
Analysis of Variance}. Cambridge University Press, Cambridge. \doi{10.1017/CBO9780511806407}.
}
\seealso{
\code{\link[AAV]{sym4asym}},
\code{\link[AAV]{asym.baci}},
\code{\link[AAV]{aav}}
}
\author{
Paulo Pagliosa \email{paulo.pagliosa@ufsc.br}
}
