% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmagnitude.R
\name{ggmagnitude}
\alias{ggmagnitude}
\title{Create a plot showing the estimated magnitude of an ecological impact}
\usage{
ggmagnitude(data, tail = "upper")
}
\arguments{
\item{data}{a list genereted from \code{\link[AAV]{magnitude}}.}

\item{tail}{only for \code{aci} post-impact design with only one disturbed location
and multiple control locations.
Perform an \code{upper} tailed, \code{lower} tailed or both join
(\code{upper_lower}) plot for all terms of the \code{aav} analyses.
Default is \code{upper}.}
}
\value{
The function returns a xy plot with the original (top right), a random (= 0; bottom left)
and the estimated alterations on the mean and variance of the data (center).
}
\description{
It provides a customised 'ggplot' for displaying the magnitude of an ecological
impact based on the results from \code{\link[AAV]{magnitude}}.
}
\details{
The magnitude of an ecological impact is estimated using the same data types, designs, factors
and parameters as in \code{\link[AAV]{aav}}, but only considers one variable at a time.

The estimates are plotted to each relevant term from the asymmetrical analysis of variance
in order to determine the magnitude of the impact.

The p-values in \code{aav} analyses came from an upper-tail test, except
wheater the analyses for \code{aci} post-impact design with only one disturbed location
are performed for a two-tailed test (i.e. aci.tailed.test = 2).
In this case, a plot showing an upper-tailed, lower-tailed or both types of tail
can be produced using the \code{tail} parameter.

The terms can be interpreted separated each other.
To each term there is a chart showing the observed p-values from the original data,
and the expected p-value from a random data (resampled the original data)
and from each estimated alteration on the mean and variance of the data.

The magnitude of the impact can be discovered when the original p-value from a term
is significant (p < 0.05). In this case, two magnitudes can be identified.
The first significant alteration to the mean and variance indicates the minimum
magnitude of impact that the original dataset can detect.
Estimates with the same p-value as the original data may represent
the magnitude of the detected impact.

Attention should be paid to whether the random data is significant.
In these cases, some spurious significant p-values may occur.

Usually, lower significance (e.g. p < 0.05) can be seen flowing from the
bottom right to the top left side of the chart.
Conversely, when visualising lower-tail tests the lower significance can be seen
flowing from the top left to the bottom right of the chart.
In the cases of spurious significance, the lower significance can be seen
flowing from the bottom left to the top right.

Wheather a term is significant, it could detect:

In baci desing:
\itemize{
\item Before vs After x Impact vs Controls = B x I:   Long-term local-scale impact
\item B x Site(I) = B x S(I):   Long-term small-scale impact
\item Perido(After) x I = P(Aft) x I:   Medium-term local-scale impact
\item P(Aft) x S(I):   Medium-term small-scale impact
\item Time(P(Aft)) x I = T(P(Aft)) x I or T(Aft) x I:   Short-term local-scale impact
\item T(P(Aft)) x S(I) or T(Aft) x S(I):   Short-term small-scale impact
}

In post-impact design with only one disturbed location and multiple control locations:
\itemize{
\item Impact vs Controls = I:   Long-term local-scale impact
\item Site(I) = S(I):   Long-term small-scale impact
\item Period x I = P x I:   Medium-term local-scale impact
\item P x S(I):   Medium-term small-scale impact
\item T(P) x I or T x I:   Short-term local-scale impact
\item T(P) x S(I) or T x S(I):   Short-term and small-scale impact
}

In post-impact design involving more than one place with one disturbed
location and multiple control locations:
\itemize{
\item Treatment = T:   Generalized local-scale impact
\item Site(T) = S(T):   Generalized site-scale impact
\item Place x T = P x T:   Local-specific scale impact
\item P x S(T):   Site-specific scale impact
}
}
\examples{
# Example for BACI design containing 2 temporal factors and 1 spatial factor.
# using the vectors: magnitude_md = c(0.10, 0.50, 1); magnitude_var = c(0.1, 1, 10)
library(ggplot2)
data(ex.ggmag)
ggmagnitude(data = ex.ggmag)

}
\references{
Underwood, A.J. 1997. \emph{Experiments in Ecology: Their Logical Design and Interpretation Using
Analysis of Variance}. Cambridge University Press, Cambridge. \doi{10.1017/CBO9780511806407}.
}
\seealso{
\code{\link[AAV]{aav}},
\code{\link[GAD]{magnitude}}
}
\author{
Paulo Pagliosa \email{paulo.pagliosa@ufsc.br}
}
