% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym4asym.R
\name{sym4asym}
\alias{sym4asym}
\title{Symmetrical ANOVAs to Run an Asymmetrical ANOVA}
\usage{
sym4asym(data, aav.design, n.ftemp, n.fspac, names.impact,
  names.before = NULL, aav.file = NULL)
}
\arguments{
\item{data}{a data frame with temporal factors, spatial factors and variable(s).}

\item{aav.design}{\code{baci} for before/after-control/impact design or \code{aci} for post-impact design.}

\item{n.ftemp}{number of temporal factors.}

\item{n.fspac}{number of spatial factors.}

\item{names.impact}{name of the asymmetrical level of the Control/Impact factor (e.g. "impact").}

\item{names.before}{only for \code{baci} design. Name of the before level of the Before/After factor (e.g. "before").}

\item{aav.file}{A file path to save the AAV output files. Default \code{NULL} uses current name and working directory.}
}
\value{
\code{Excel} files containing the symmetrical ANOVA tables used to
run the asymmetrical ANOVAs.
When analyzing more than one variable at once, each sheet have one ANOVA table.
}
\description{
Calculate the symmetrical ANOVAs to run an asymmetrical ANOVA
beyond BACI or post-impact design.

Asymmetrical ANOVAs are essentially multi-factorial.
Beyond BACI design can be used for models with 2 or 3 temporal factors
and 1 or 2 spatial factors. Post-impact design with only one disturbed location
and multiple control locations can be used for models with 1 or 2 temporal factors
and 1 or 2 spatial factors. Post-impact design with more than one place with
one disturbed location and multiple control locations can be used for models
with 2 or 3 spatial factors.
}
\details{
The construction of asymmetrical analyses of variance consists of combining
the Sum of Squares values from separate (symmetrical) ANOVAs.
In beyond BACI designs, four fully orthogonal ANOVAs need to be done: on all
the data, on "control" data, on "before" data, and on "control" and "before" data.
In post-impact designs with only one disturbed location two fully orthogonal
ANOVAs need to be done: on all the data and on "control" data.
In post-impact designs with more than one place with one disturbed location
two fully orthogonal and one fully nested ANOVAs need to be done:
on all data (orthogonal) and on "control" data (orthogonal and nested).
\code{sym4asy} use \code{\link[GAD]{GAD}} (Sandrini-Neto et al. 2024) to perform symmetrical ANOVAs.

The \code{sym4asym} analysis fits mixed ANOVA designs with a specific
combination of orthogonal/nested and fixed/random factors:
The  beyond BACI design runs for five-factors (3 temporal and 2 spatial),
four-factors (3 temporal and 1 spatial or 2 temporal and 2 spatial), and
three-factors (2 temporal and 1 spatial).
The post-impact design with only one disturbed location runs for
four-factors (2 temporal and 2 spatial), three-factors (2 temporal and 1 spatial or
1 temporal and 2 spatial), and two-factors (1 temporal and 1 spatial).
The post-impact design with more than one place with one disturbed location runs for
three-factors (0 temporal and 3 spatial) and two-factors (0 temporal and 2 spatial).

In all models the Location factor is the asymmetric one (e.g., one "impact"
group and two or more "control" groups), meaning there are different numbers
of levels of this nested factor within each different level of its upper-level
factor (e.g., Before vs After, Period, Time or Place factors).

In all models the temporal and spatial factors are orthogonal.
Within each factor type (temporal or spatial) the factors are hierarchically nested.
In the beyond BACI and post-impact design with only one disturbed location models,
the first factor and the Location factor are fixed and all others are random.
In the post-impact design with more than one place with one disturbed location,
the first factor is fixed and all other factors are random.

The data frame must follow a specific structure:
1- first the temporal factors, then the spatial factors, and then the variable(s);
2- each factor type must appear in a hierarchical order.

For beyond BACI design the order for temporal factors is Before_after, Period and Time (3 factors)
or Before_after and Time (2 factors).
The order for spatial factors is Location and Site (2 factors)
or just Location (1 factor).
For post-impact design with only one disturbed location and multiple control locations,
the order for temporal factors must be Period and Time (2 factors) or just Time (1 factor), and
the order for spatial factors must be Location and Site (2 factors) or just Location (1 factor).
For post-impact design with more than one place with one disturbed location and multiple control locations,
the order for spatial factors must be Place, Location and Site (3 factors) or Place and Location (2 factors).

If the names of the factors are different from the above names they will be changed.

When the data frame have more than one variable the univariate ANOVAs are run
for each one separated.
}
\examples{
# Example for BACI design containing 3 temporal factors and 2 spatial factor:
data(baci.3t2s)
ex.baci.3t2s<-sym4asym(data = baci.3t2s, aav.design = "baci", n.ftemp = 3, n.fspac = 2,
names.impact = "impact", names.before = "before")
ex.baci.3t2s

# Example for post-impact design with only one disturbed location
# containing 2 temporal factors and 2 spatial factor:
data(aci.2t2s)
ex.aci.2t2s<-sym4asym(data = aci.2t2s, aav.design = "aci", n.ftemp = 2, n.fspac = 2,
    names.impact = "impact")
ex.aci.2t2s

# Example for post-impact design with more than one place with one
# disturbed location containing 0 temporal factors and 3 spatial factor:
 data(aci.0t3s)
ex.aci.0t3s<-sym4asym(data = aci.0t3s, aav.design = "aci", n.ftemp = 0, n.fspac = 3,
    names.impact = "impact")
ex.aci.0t3s

}
\references{
Glasby, T.M. 1997. Analysing data from post-impact studies using asymmetrical analyses
of variance: A case study of epibiota on marinas. Aust. J. Ecol., 22, 448-459.
\doi{10.1111/j.1442-9993.1997.tb00696.x}.

Sandrini-Neto, L., Gilbert, E., Camargo, M. 2024. \emph{GAD: Analysis of Variance from General Principles}.
R package version 2.0, \doi{10.32614/CRAN.package.GAD}.

Underwood, A.J. 1997. \emph{Experiments in Ecology: Their Logical Design and Interpretation Using
Analysis of Variance}. Cambridge University Press, Cambridge. \doi{10.1017/CBO9780511806407}.
}
\seealso{
\code{\link[AAV]{aav}},
\code{\link[AAV]{asym.aci}},
\code{\link[AAV]{asym.baci}},
\code{\link[GAD]{gad}}
}
\author{
Paulo Pagliosa \email{paulo.pagliosa@ufsc.br}
}
